data-pipeline/requirements.txt
httpx
supabase
python-dotenv
apscheduler
pandas
